# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: frequency
          schema:
            type: string
            enum: [monthly, annually, one_time]
          required: false
        - in: query
          name: status
          schema:
            type: string
            enum: [active, cancelled]
          required: false
        - in: query
          name: from
          schema:
            type: string
          required: false
        - in: query
          name: to
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/summary:
    get:
      operationId: getExpenseSummary
      tags: [expenses]
      summary: Get expense totals grouped by frequency and category
      parameters:
        - in: query
          name: from
          schema:
            type: string
          required: false
        - in: query
          name: to
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Summary totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseSummary"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get a single expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense permanently
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}/cancel:
    patch:
      operationId: cancelExpense
      tags: [expenses]
      summary: Cancel (deactivate) an expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}/receipt:
    post:
      operationId: generateReceipt
      tags: [receipts]
      summary: Generate a receipt for an expense and email it
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt generated and emailed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /financials/overview:
    get:
      operationId: getFinancialOverview
      tags: [financials]
      summary: Get P&L overview comparing money in vs money out
      parameters:
        - in: query
          name: from
          schema:
            type: string
          required: false
        - in: query
          name: to
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Financial overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialOverview"
  /financials/statement:
    get:
      operationId: getAccountStatement
      tags: [financials]
      summary: Get account statement with all transactions
      parameters:
        - in: query
          name: from
          schema:
            type: string
          required: false
        - in: query
          name: to
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Account statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStatement"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Expense:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        frequency:
          type: string
          enum: [monthly, annually, one_time]
        status:
          type: string
          enum: [active, cancelled]
        notes:
          type: ["string", "null"]
        renewalDate:
          type: ["string", "null"]
        purchaseDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - amount
        - currency
        - category
        - frequency
        - status
        - notes
        - renewalDate
        - purchaseDate
        - createdAt
        - updatedAt
    CreateExpenseBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        frequency:
          type: string
          enum: [monthly, annually, one_time]
        notes:
          type: ["string", "null"]
        renewalDate:
          type: ["string", "null"]
        purchaseDate:
          type: ["string", "null"]
      required:
        - name
        - amount
        - category
        - frequency
    UpdateExpenseBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        frequency:
          type: string
          enum: [monthly, annually, one_time]
        status:
          type: string
          enum: [active, cancelled]
        notes:
          type: ["string", "null"]
        renewalDate:
          type: ["string", "null"]
        purchaseDate:
          type: ["string", "null"]
    ExpenseSummary:
      type: object
      properties:
        totalMonthly:
          type: number
        totalAnnually:
          type: number
        totalOneTime:
          type: number
        totalAllAnnualized:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"
        activeCount:
          type: integer
        cancelledCount:
          type: integer
      required:
        - totalMonthly
        - totalAnnually
        - totalOneTime
        - totalAllAnnualized
        - byCategory
        - activeCount
        - cancelledCount
    CategorySummary:
      type: object
      properties:
        category:
          type: string
        total:
          type: number
        count:
          type: integer
      required:
        - category
        - total
        - count
    Receipt:
      type: object
      properties:
        id:
          type: integer
        expenseId:
          type: integer
        receiptNumber:
          type: string
        emailedTo:
          type: string
        generatedAt:
          type: string
        expenseName:
          type: string
        expenseAmount:
          type: number
        expenseCategory:
          type: string
        expenseFrequency:
          type: string
      required:
        - id
        - expenseId
        - receiptNumber
        - emailedTo
        - generatedAt
        - expenseName
        - expenseAmount
        - expenseCategory
        - expenseFrequency
    ReceiptResponse:
      type: object
      properties:
        receipt:
          $ref: "#/components/schemas/Receipt"
        message:
          type: string
        emailSent:
          type: boolean
      required:
        - receipt
        - message
        - emailSent
...
    FinancialOverview:
      type: object
      properties:
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netPL:
          type: number
        isInRed:
          type: boolean
        annualizedIncome:
          type: number
        annualizedExpenses:
          type: number
        annualizedNetPL:
          type: number
        monthlyIncome:
          type: number
        monthlyExpenses:
          type: number
        monthlyNetPL:
          type: number
        incomeByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"
        expensesByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"
      required:
        - totalIncome
        - totalExpenses
        - netPL
        - isInRed
        - annualizedIncome
        - annualizedExpenses
        - annualizedNetPL
        - monthlyIncome
        - monthlyExpenses
        - monthlyNetPL
        - incomeByCategory
        - expensesByCategory
    StatementEntry:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        name:
          type: string
        amount:
          type: number
        currency:
          type: string
        category:
          type: string
        frequency:
          type: string
        status:
          type: string
        date:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - name
        - amount
        - currency
        - category
        - frequency
        - status
        - date
        - createdAt
    AccountStatement:
      type: object
      properties:
        from:
          type: ["string", "null"]
        to:
          type: ["string", "null"]
        generatedAt:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/StatementEntry"
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netPL:
          type: number
      required:
        - from
        - to
        - generatedAt
        - entries
        - totalIncome
        - totalExpenses
        - netPL
...