# lib/api-zod/src/generated/types/createExpenseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExpenseBodyFrequency } from "./createExpenseBodyFrequency";

export interface CreateExpenseBody {
  name: string;
  amount: number;
  currency?: string;
  category: string;
  frequency: CreateExpenseBodyFrequency;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  renewalDate?: string | null;
  /** @nullable */
  purchaseDate?: string | null;
}