# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseFrequency } from "./expenseFrequency";
import type { ExpenseStatus } from "./expenseStatus";

export interface Expense {
  id: number;
  name: string;
  amount: number;
  currency: string;
  category: string;
  frequency: ExpenseFrequency;
  status: ExpenseStatus;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  renewalDate: string | null;
  /** @nullable */
  purchaseDate: string | null;
  createdAt: string;
  updatedAt: string;
}