# lib/api-zod/src/generated/types/statementEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatementEntryType } from "./statementEntryType";

export interface StatementEntry {
  id: number;
  type: StatementEntryType;
  name: string;
  amount: number;
  currency: string;
  category: string;
  frequency: string;
  status: string;
  /** @nullable */
  date: string | null;
  createdAt: string;
}